+++ src/main/java/ma/vi/esql/semantic/type/BaseType.java
@@ -42,7 +42,11 @@ public class BaseType extends AbstractType {
   }
 
   @Override
-  public String translate(Target target, EsqlConnection esqlCon, EsqlPath path, PMap<String, Object> parameters, Environment env) {
+  public String translate(Target               target,
+                          EsqlConnection       esqlCon,
+                          EsqlPath             path,
+                          PMap<String, Object> parameters,
+                          Environment          env) {
     return target == Target.ESQL            ? name() :
            translations.containsKey(target) ? translations.get(target) : translations.get(Target.ALL);
   }